Connected new Start Button and Don't Lose Progress page
