+++ packages/web-app/src/modules/machine-views/DontLoseProgressContainer.ts
@@ -1,0 +1,13 @@
+import { connect } from '../../connect'
+import { RootStore } from '../../Store'
+import { StartActionType } from '../salad-bowl/models'
+import { DontLoseProgressPage } from './components/DontLoseProgress'
+
+const mapStoreToProps = (store: RootStore): any => ({
+  onCloseClicked: () => store.ui.hideModal(true),
+  onSubmitSupportTicket: () => store.zendesk.openSupportTicket(),
+  onStopPrepping: () => store.saladBowl.toggleRunning(StartActionType.StopPrepping),
+  prepTime: store.saladBowl.runningTime,
+})
+
+export const DontLoseProgressPageContainer = connect(mapStoreToProps, DontLoseProgressPage)
