+++ packages/web-app/src/modules/machine-views/StartButtonContainer.ts
@@ -1,16 +1,17 @@
 import { connect } from '../../connect'
 import { RootStore } from '../../Store'
+import { ErrorPageType } from '../../UIStore'
 import { StartActionType } from '../salad-bowl/models'
 import { StartButton } from './components/StartButton'
 
 const mapStoreToProps = (store: RootStore): any => ({
-  onClick: () => store.saladBowl.toggleRunning(StartActionType.StartButton),
+  isLarge: true,
   isRunning: store.saladBowl.isRunning,
-  status: store.saladBowl.status,
-  notCompatible: store.saladBowl.isNotCompatible,
+  onClick: () => store.saladBowl.toggleRunning(StartActionType.StartButton),
+  onClickError: store.saladBowl.isNotCompatible ? () => store.ui.showErrorPage(ErrorPageType.NotCompatible) : undefined,
+  progress: store.saladBowl.preppingProgress,
   runningTime: store.saladBowl.runningTime,
-  pluginCount: store.saladBowl.pluginCount,
-  isDesktop: store.native.isNative,
+  status: store.saladBowl.status,
 })
 
 export const StartButtonContainer = connect(mapStoreToProps, StartButton)
