+++ packages/web-app/src/modules/machine-views/index.ts
@@ -1,3 +1,4 @@
+export * from './DontLoseProgressContainer'
 export * from './MachineInfoContainer'
 export * from './MinerListContainer'
 export * from './MinerTypeContainer'
