+++ packages/web-app/src/modules/salad-bowl/models/StartActionType.ts
@@ -2,5 +2,7 @@ export enum StartActionType {
   Automatic = 'Automatic',
   Override = 'Override',
   StartButton = 'Start Button',
+  StopPrepping = 'Stop Prepping',
   SwitchMiner = 'Switch Miner',
+  TimeToChopStartButton = 'Time To Chop Start Button',
 }
